+Job&Person class i deres egen fil +List i library
